use parseFeedFromFile to avoid mojibake
As mentioned in commit 2bd778a46ed071c2dd534ebe3b7007b7ae60d1c1, there
was mojibake when LANG=C.

Looking at parseFeedFromFile, it is very particular to read the file as
unicode. parseFeedString looks like it will accept any old String,
but a String that was read using the filesystem encoding will not in
fact have the right encoding.

I think this is a bug in the feed library and will file one.

Sponsored-by: Svenne Krap on Patreon