fix too early close of shared lock file
This fixes a reversion introduced in commit
ac56a5c2a056d3df1314a0a0cc16d276ddefb661.

I didn't notice there that it was handling the case of a shared lock
file that was still open elsewhere by not running the close action.

This was especially deadly when annex.pidlock is set, as it caused early
deletion of the pid lock file.

Sponsored-by: Dartmouth College's Datalad project