Revert "fix too early close of shared lock file"
This reverts commit 66b2536ea0aa5c88f5b744eeace3322a8a4a10b6.

I misunderstood commit ac56a5c2a056d3df1314a0a0cc16d276ddefb661
and caused a FD leak when pid locking is not used.

A LockHandle contains an action that will close the underlying lock
file, and that action is run when it is closed. In the case of a shared
lock, the lock file is opened once for each LockHandle, and only
the one for the LockHandle that is being closed will be closed.